+++ pcdmis_toolbox/modules/pc_to_excel/inject/__init__.py
@@ -4,10 +4,12 @@ from .command_injector import (
     generate_command_text,
     inject_export_command,
 )
+from .toolbar_launcher import deploy_toolbar_launcher
 
 __all__ = [
     "check_export_command",
     "deploy_bas_script",
+    "deploy_toolbar_launcher",
     "generate_command_text",
     "inject_export_command",
 ]